Struct ori_font_awesome::Icon

pub struct Icon {
    pub icon: IconCode,
    pub solid: bool,
    pub size: Styled<f32>,
    pub color: Styled<Color>,
}

A view that displays a single icon.

By default, the icon is rendered using the icon.font font family. This uses the Font Awesome 6 Regular Free font by default.

Fields
icon: IconCode

The codepoint of the icon to display.

solid: bool

Whether the icon is solid or regular.

This only affects the rendering of the icon if the icon is available in both styles.

size: Styled<f32>

The size of the icon.

color: Styled<Color>

The color of the icon.

Implementations

impl Icon

pub fn icon(self, icon: impl Into<IconCode>) -> Self

Set self.icon.

The codepoint of the icon to display.

pub fn solid(self, solid: impl Into<bool>) -> Self

Set self.solid.

Whether the icon is solid or regular.

This only affects the rendering of the icon if the icon is available in both styles.

pub fn size(self, size: impl Into<Styled<f32>>) -> Self

Set self.size.

The size of the icon.

pub fn color(self, color: impl Into<Styled<Color>>) -> Self

Set self.color.

The color of the icon.

impl Icon

pub fn new(icon: impl Into<IconCode>) -> Self

Create a new icon view.

pub fn font(&self) -> IconFont

Get the font to use for the icon.